==> 00. Archive/03. Glacier simulations/02. r1_a10/03. Run OGGM w. custom climate data - sample glaciers.py <==
""" This script runs the OGGM step by step guide to a preprocessed directory"""# %% Cell 1: Load the required directories# from OGGM_data_processing import process_perturbation_datafrom OGGM_data_processing import process_perturbation_dataimport oggmfrom oggm import cfg, utils, workflow, tasks, DEFAULT_BASE_URLimport geopandas as gpdimport numpy as npimport osfrom oggm.shop import rgitopofrom oggm import GlacierDirectoryfrom oggm.core import massbalance, flowline# Libsimport xarray as xrimport matplotlib.pyplot as pltimport matplotlib.cm as cmfrom oggm.shop import cru, w5e5from oggm.tasks import process_w5e5_datafrom oggm.core.massbalance import mb_calibration_from_geodetic_mbimport matplotlib.pyplot as pltfrom oggm import graphicsimport pandas as pdimport loggingimport shutilimport argparseimport timeimport loggingimport jsonimport csvimport sys# %% Cell 1: Create a functio to add new files to basenamesdef add_to_basenames(basename, filename, docstr=''):    """Add an entry to the list of BASENAMES.    BASENAMES are access keys to files available at the gdir level.    Parameters    ----------    basename : str        the key (e.g. 'dem', 'model_flowlines')    filename : str        the associated filename (e.g. 'dem.tif', 'model_flowlines.pkl')    docstr : str        the associated docstring (for documentation)    """    global BASENAMES    if '.' not in filename:        raise ValueError('The filename needs a proper file suffix!')    cfg.BASENAMES[basename] = (filename, docstr)# cfg.add_to_basenames('model_flowlines_dyn_melt_f_calib.pkl', 'model_flowlines.pkl', 'Adding the dynamic melt calibration flowlines')# %% Cell 2: Provide modelling parameters for OGGM# Use W5E5 data# folder_path='/Users/magaliponds/Library/CloudStorage/OneDrive-VrijeUniversiteitBrussel/1. VUB/02. Coding/01. IRRMIP/03. Data/05. OGGM Working Directories'folder_path = '/Users/magaliponds/Documents/00. Programming'wd_path = f'{folder_path}/01. Modelled perturbation-glacier interactions/'y0_clim = 1985ye_clim = 2014y0_spinup = 1980ye_spinup = 2020colors = {    "W5E5": ["#000000"],  # "#000000"],  # Black    # Darker to lighter shades of purple    "E3SM": ["#785EF0", "#8F7BF1", "#A6A8F2"],    # Darker to lighter shades of pink    "CESM2": ["#DC267F", "#E58A9E", "#F0A2B6", "#F7BCC4"],    # Darker to lighter shades of orange    "CNRM": ["#FE6100", "#FE7D33", "#FE9A66", "#FEB799", "#FECDB5", "#FEF1E1"],    "IPSL-CM6": ["#FFB000"]  # Dark purple to lighter shades}# Specify the RGI region you are interested in, for example, region 13 (Central Asia)rgi_reg = '13'  # this must fit to example glacier(s), if starting from level 0rgi_ids = ["RGI60-13.40102",  # Area >10km2           "RGI60-13.39195",  # Area >10km2           "RGI60-13.36881",  # Area >10km2           "RGI60-13.38969",  # Area >10km2           "RGI60-13.37184",  # Area >10km2           "RGI60-13.00967",  # Area <10km2|           "RGI60-13.40982",  # Area <10km2           "RGI60-13.41891",  # WGMS observation           "RGI60-13.23659",  # WGMS observation           ]base_url_elev = (    'https://cluster.klima.uni-bremen.de/~oggm/gdirs/oggm_v1.6/L3-L5_files/2023.3/elev_bands/W5E5/')base_url_center = (    'https://cluster.klima.uni-bremen.de/~oggm/gdirs/oggm_v1.6/L3-L5_files/2023.3/centerlines/W5E5/')prepro_level = 4border = 80# fill out if gdirs should be resetted and gdirs should be downloadedreset = Falsedownload_gdirs = Falsecfg.initialize(logging_level='WARNING')cfg.PARAMS['baseline_climate'] = "GSWP3-W5E5"cfg.PATHS['working_dir'] = utils.mkdir(wd_path, reset=reset)sum_dir = os.path.join(wd_path, 'summary')os.makedirs(sum_dir, exist_ok=True)cfg.PARAMS['use_multiprocessing'] = Falsecfg.PARAMS['store_model_geometry'] = Trueif download_gdirs == True:    gdirs = workflow.init_glacier_directories(rgi_ids,                                              from_prepro_level=4,                                              prepro_base_url=base_url_elev, prepro_border=border)else:    gdirs = workflow.init_glacier_directories(rgi_ids)# %% Cell 0: Process the perturbationsfunction_directory = "/Users/magaliponds/Library/CloudStorage/OneDrive-VrijeUniversiteitBrussel/1. VUB/02. Coding/01. IRRMIP/src/03. Glacier simulations"sys.path.append(function_directory)members = [1, 3, 4, 6, 1]models = ["IPSL-CM6", "E3SM", "CESM2", "CNRM"]timeframe = "monthly"opath_climate = os.path.join(sum_dir, 'climate_historical.nc')utils.compile_climate_input(    gdirs, path=opath_climate, filename='climate_historical')for m, model in enumerate(models):    for member in range(members[m]):        # Provide the path to the perturbation dataset        o_folder_ptb = f"/Users/magaliponds/Library/CloudStorage/OneDrive-VrijeUniversiteitBrussel/1. VUB/02. Coding/01. IRRMIP/03. Data/03. Output files/01. Climate data/07. OGGM Perturbation input files/{timeframe}/{model}/{member}"        ds_path = f"{o_folder_ptb}/{model}.00{member}.{y0_clim}_{ye_clim}.{timeframe}.perturbation.input.%.degC.nc"        # Provide the sample ID to provide the processed pertrubations with the correct output suffix        sample_id = f"{model}.00{member}"        workflow.execute_entity_task(process_perturbation_data, gdirs,                                     ds_path=ds_path,                                     y0=None, y1=None,                                     output_filesuffix=f'_perturbation_{sample_id}')        opath_perturbations = os.path.join(            sum_dir, f'climate_historical_perturbation_{sample_id}.nc')        utils.compile_climate_input(gdirs, path=opath_perturbations, filename='climate_historical',                                    input_filesuffix=f'_perturbation_{sample_id}',                                    use_compression=True)# %% Cell 1: Inspect the output files - do they have the same DEMopath_climate = os.path.join(sum_dir, 'climate_historical.nc')members = [1, 3, 4, 6, 1]models = ["IPSL-CM6", "E3SM", "CESM2", "CNRM"]timeframe = "monthly"# rgi_id = gdirs[0].rgi_idplt.figure()size = 16for m, model in enumerate(models):    size = size-2    for member in range(members[m]):        opath_perturbations = os.path.join(            sum_dir, f'climate_historical_perturbation_{sample_id}.nc')        ds_ptb = xr.open_dataset(opath_perturbations)        plt.scatter(ds_ptb.rgi_id, ds_ptb.ref_hgt,                    color=colors[model][member], s=size, label=f"{model}.00{member}")ds_base = xr.open_dataset(opath_climate)plt.scatter(ds_base.rgi_id, ds_base.ref_hgt, color='k',            alpha=1, label='W5E5.000 baseline', s=8)plt.xticks(rotation=45)plt.ylabel("Reference Height [m]")plt.xlabel("RGI ID")plt.legend(bbox_to_anchor=(0.5, -0.7), loc='lower center', ncols=4)# %% Cell 2: Create a plot for the perturbations vs baseline climatecolors = {    "W5E5": ["#000000"],  # "#000000"],  # Black    # Darker to lighter shades of purple    "E3SM": ["#785EF0", "#8F7BF1", "#A6A8F2"],    # Darker to lighter shades of pink    "CESM2": ["#DC267F", "#E58A9E", "#F0A2B6", "#F7BCC4"],    # Darker to lighter shades of orange    "CNRM": ["#FE6100", "#FE7D33", "#FE9A66", "#FEB799", "#FECDB5", "#FEF1E1"],    "IPSL-CM6": ["#FFB000"]  # Dark purple to lighter shades}# Initialize figure for temperature and precipitationfig_temp, axes_temp = plt.subplots(    3, 3, figsize=(15, 8))  # 3x3 grid for temperaturefig_prcp, axes_prcp = plt.subplots(    3, 3, figsize=(15, 8))  # 3x3 grid for precipitation# Flatten axes for easy iterationaxes_temp = axes_temp.flatten()axes_prcp = axes_prcp.flatten()for idx, gdir in enumerate(gdirs):    tasks.init_present_time_glacier(gdir)    # Load historical climate data    with xr.open_dataset(gdir.get_filepath('climate_historical')) as ds:        ds = ds.load()    ds = ds.sel(time=slice("1985-01-01", "2014-12-01"))    ds = ds.resample(time='YS').mean()    # Create primary axes for baseline climate data    ax_temp = axes_temp[idx]    ax_prcp = axes_prcp[idx]    # Create twin axes only once per subplot    ax_temp_perturb = ax_temp.twinx()    ax_prcp_perturb = ax_prcp.twinx()    # Plot settings    members = [1, 3, 4, 6, 1]    models = ["IPSL-CM6", "E3SM", "CESM2", "CNRM"]    # Loop over models and members    for m, model in enumerate(models):        for member in range(members[m]):            sample_id = f"{model}.00{member}"            clim_ptb = ds.copy()  # Copy historical climate data            # Load perturbation data            with xr.open_dataset(gdir.get_filepath('climate_historical', filesuffix='_perturbation_{}'.format(sample_id))) as ds_ptb:                ds_ptb = ds_ptb.load()            ds_ptb = ds_ptb.sel(time=slice("1985-01-01", "2014-12-01"))            ds_ptb = ds_ptb.resample(time='YS').mean()            # Apply perturbation to temperature and precipitation            clim_ptb['temp'] = clim_ptb.temp - ds_ptb.temp            clim_ptb['prcp'] = clim_ptb.prcp - clim_ptb.prcp * ds_ptb.prcp            # Plot Temperature            ax_temp.plot(ds.time, ds.temp, 'k-',                         label='Baseline')  # Baseline climate            ax_temp.set_ylabel('Temperature (°C)', color='k')            ax_temp.tick_params('y', colors='k')            if member == 0:                # ax_temp.plot(ds_ptb.time, ds_ptb.temp, color=colors[model][member], label='Perturbation')                # ax_temp.set_ylabel('Temperature (Perturbation)', color='k')                # ax_temp.tick_params('y', colors='k')                ax_temp_perturb.plot(                    ds_ptb.time, ds_ptb.temp, color=colors[model][member], label='Perturbation')                ax_temp_perturb.set_ylabel(                    '$\Delta$ Temperature (°C)', color=colors[model][member])                ax_temp_perturb.tick_params('y', color=colors[model][member])            ax_temp_perturb.set_ylim(-1, 1)            ax_temp.set_title(f'{gdir.rgi_id}')            # Plot Precipitation            ax_prcp.plot(ds.time, ds.prcp, 'k-',                         label='Baseline')  # Baseline climate            ax_prcp.set_ylabel('$\Delta$ Precipitation (Baseline)', color='k')            ax_prcp.tick_params('y', colors='k')            if member == 0:                # ax_prcp.plot(ds_ptb.time, ds_ptb.prcp, color=colors[model][member], label='Perturbation')                # ax_prcp.set_ylabel('Precipitation (Perturbation)', color='k')                # ax_prcp.tick_params('y', colors='k')                ax_prcp_perturb.plot(                    ds_ptb.time, ds_ptb.prcp*100, color=colors[model][member], label='Perturbation')                ax_prcp_perturb.set_ylabel(                    'Precipitation (%)', color=colors[model][member])                ax_prcp_perturb.tick_params('y', colors=colors[model][member])            ax_prcp_perturb.set_ylim(-50, 50)            ax_prcp.set_title(f'{gdir.rgi_id}')            # Save the perturbed dataset            clim_ptb.to_netcdf(gdir.get_filepath(                'gcm_data', filesuffix='_perturbed_{}'.format(sample_id)))# Adjust layout and show the figuresfig_temp.tight_layout()fig_prcp.tight_layout()plt.show()# %% Cell 3: Perturb the climate data in a gcm file# gdir = gdirs[0]for gdir in gdirs:    tasks.init_present_time_glacier(gdir)    with xr.open_dataset(gdir.get_filepath('climate_historical')) as ds:        ds = ds.load()    members = [1, 3, 4, 6, 1]    models = ["IPSL-CM6", "E3SM", "CESM2", "CNRM"]    for m, model in enumerate(models):        for member in range(members[m]):            sample_id = f"{model}.00{member}"            # make a copy of the historical climate            clim_ptb = ds.copy()            # open the perturbation dataset and add the perturbations            with xr.open_dataset(gdir.get_filepath('climate_historical', filesuffix='_perturbation_{}'.format(sample_id))) as ds_ptb:                ds_ptb = ds_ptb.load()            clim_ptb['temp'] = clim_ptb.temp - ds_ptb.temp            clim_ptb['prcp'] = clim_ptb.prcp - clim_ptb.prcp * ds_ptb.prcp            clim_ptb = clim_ptb.dropna('time')            clim_ptb.to_netcdf(gdir.get_filepath(                'gcm_data', filesuffix='_perturbed_{}'.format(sample_id)))# %% Cell 4: Open output datasetwith xr.open_dataset(gdir.get_filepath('gcm_data', filesuffix='_perturbed_{}'.format(sample_id))) as ds_ptb2:    ds_ptb2 = ds_ptb2.load()print(ds_ptb2.time)# %% Cell 5: Uses the MB model to calculate specific mass balances for each of the perturbed climates# download the geodetic MB dataframemb_geodetic = utils.get_geodetic_mb_dataframe()# set up a dataframe for mbyears = np.arange(1985, 2014)years_base = np.arange(1985, 2020)mb = pd.DataFrame({'years': years_base})mb['years'] = pd.to_datetime(mb['years'], format='%Y')gdirs = workflow.init_glacier_directories(rgi_ids)members = [1, 3, 4, 6, 1]models = ["IPSL-CM6", "E3SM", "CESM2", "CNRM"]fig, axes = plt.subplots(3, 3, figsize=(15, 8), sharex=True)for (g, gdir) in enumerate(gdirs):    row = g // axes.shape[1]    col = g % axes.shape[1]    # Get the calibrated mass-balance model - the default is to use OGGM's "MonthlyTIModel"    mbmod = massbalance.MonthlyTIModel(gdir)    # Compute the specific MB for this glacier    fls = gdir.read_pickle('model_flowlines')    mb_ts = mbmod.get_specific_mb(fls=fls, year=years_base)    # add the outcome to the dataframe and calculate mean values for the provided time slices    mb['B'] = mb_ts    mb_2020 = mb[(mb['years'] >= '2000-01-01') & (mb['years'] < '2020-01-01')]    mean_mb_2020 = mb_2020['B'].mean()    # load the geodetic mass balance    sel = mb_geodetic.loc[gdir.rgi_id].set_index('period') * 1000    # download the mb and error for Hugonnet    _mb, _err = sel.loc['2000-01-01_2020-01-01'][['dmdtda', 'err_dmdtda']]    axes[row, col].fill_between(        [2000, 2020], [_mb-_err, _mb-_err], [_mb+_err, _mb+_err], alpha=0.5, color='grey')    axes[row, col].plot([2000, 2020], [_mb, _mb],                        color='grey', label='Hugonnet 2000-2020')    for m, model in enumerate(models):        timeseries_data = []        for member in range(members[m]):            sample_id = f"{model}.00{member}"            mbmod2 = massbalance.MonthlyTIModel(                gdir, filename='gcm_data', input_filesuffix='_perturbed_{}'.format(sample_id))            mb_ts2 = mbmod2.get_specific_mb(fls=fls, year=years)            timeseries_data.append(mb_ts2)            if member == 0:                axes[row, col].plot(                    years, mb_ts2, color=colors[model][member], label=sample_id)            else:                axes[row, col].plot(                    years, mb_ts2, color=colors[model][member], label=sample_id, ls=":")        if len(timeseries_data) > 0:            stacked_data = np.stack(timeseries_data)            # print(stacked_data)            min_values = np.min(stacked_data, axis=0).flatten()            max_values = np.max(stacked_data, axis=0).flatten()        if model == "IPSL-CM6":            axes[row, col].fill_between(years, min_values, max_values, color=colors[model]                                        [0], alpha=0.3, label=f"{model}, {members[m]}-member range")        else:            axes[row, col].fill_between(years, min_values, max_values, color=colors[model]                                        [0], alpha=0.3, label=f"{model}, {members[m]-1}-member range")    # plot baseline    axes[row, col].plot([2000, 2020], [mean_mb_2020, mean_mb_2020],                        color=colors["W5E5"][0], linestyle='dashed', label='W5E5.000 modelled mean 2000-2014')    axes[row, col].plot(years_base, mb.B, color=colors["W5E5"]                        [0], label='W5E5.000 baseline')    axes[row, col].set_title(gdir.rgi_id)    if g == 3:        axes[row, col].set_ylabel('Specific MB (mm w.e.)', fontsize=12)    if g == 7:        axes[row, col].set_xlabel('Year', fontsize=12)    plt.legend(ncols=4, loc="lower center", bbox_to_anchor=(-0.8, -1.2))# %% Cell 6: Run the climate model from GCM data# gdir = gdirs[0]y0_clim = 1985ye_clim = 2014members = [1, 3, 4, 6, 1]models = ["IPSL-CM6", "E3SM", "CESM2", "CNRM"]for m, model in enumerate(models):    for member in range(members[m]):        sample_id = f"{model}.00{member}"        # for gdir in gdirs:        workflow.execute_entity_task(tasks.init_present_time_glacier, gdirs)        # ipath = os.path.join(        #     wd_path, "per_glacier", rgi_id[:-6], rgi_id[:-3], rgi_id, f'gcm_data_perturbed_{sample_id}.nc')        # with xr.open_dataset(ipath) as ds:        #     ds = ds.load()        # print(ds)        out_id = f'_perturbed_{sample_id}'        opath = os.path.join(            sum_dir, f'climate_run_output_perturbed_{sample_id}.nc')        workflow.execute_entity_task(tasks.run_from_climate_data, gdirs,                                     ys=y0_clim, ye=ye_clim,  # min_ys=None,                                     max_ys=None, fixed_geometry_spinup_yr=None,                                     store_monthly_step=False, store_model_geometry=None,                                     store_fl_diagnostics=True, climate_filename='gcm_data',                                     # mb_model=None, mb_model_class=<class 'oggm.core.massbalance.MonthlyTIModel'>,                                     climate_input_filesuffix='_perturbed_{}'.format(                                         sample_id),                                     output_filesuffix=out_id,                                     # init_model_filesuffix='_historical', #in case you want to start from a previous model state                                     # init_model_yr=2015, init_model_fls=None,                                     zero_initial_glacier=False, bias=0,                                     temperature_bias=None, precipitation_factor=None)        ds_ptb = utils.compile_run_output(            gdirs, input_filesuffix=out_id, path=opath)# And run the climate model with reference dataworkflow.execute_entity_task(tasks.run_from_climate_data, gdirs,                             ys=y0_clim, ye=ye_clim,                             output_filesuffix='_baseline_W5E5.000')opath_base = os.path.join(sum_dir, 'climate_run_output_baseline_W5E5.000.nc')ds_base = utils.compile_run_output(    gdirs, input_filesuffix='_baseline_W5E5.000', path=opath_base)# %% Cell 8: Plot volume and area evolutiondef plot_volume_area_evolution_over_time_combined(averaged):    timeframe = "monthly"    members = [3, 4, 6, 1, 1]    models = ["E3SM", "CESM2", "CNRM", "IPSL-CM6", "W5E5"]    variables = ["volume", "area"]  # "volume",    factors = [10**-9, 10**-6]    variable_names = ["Volume", "Area"]    variable_axes = ["Volume [km$^3$]", "Area [km$^2$]"]    colors = {        "W5E5": ["#000000"],        "E3SM": ["#785EF0", "#8F7BF1", "#A6A8F2"],        "CESM2": ["#DC267F", "#E58A9E", "#F0A2B6", "#F7BCC4"],        "CNRM": ["#FE6100", "#FE7D33", "#FE9A66", "#FEB799", "#FECDB5", "#FEF1E1"],        "IPSL-CM6": ["#FFB000"]    }    for v, var in enumerate(variables):        fig, axes = plt.subplots(3, 3, figsize=(15, 8))        axes = axes.flatten()  # Flatten the axes array to simplify indexing        for r, rgi_id in enumerate(rgi_ids):            for m, model in enumerate(models):                timeseries_data = []                if averaged == False:                    members = members                else:                    members = [1, 1, 1, 1, 1]                for i in range(members[m]):                    sample_id = f"{model}.00{i}"                    if model != "W5E5":                        opath = os.path.join(                            sum_dir, f'climate_run_output_perturbed_{sample_id}.nc')                    else:                        opath = os.path.join(                            wd_path, "summary", f"climate_run_output_baseline_{sample_id}.nc")                    ds_diag_ext = xr.open_dataset(opath)                    # Select the correct subplot for the current rgi_id                    ax = axes[r]                    ax.set_title(rgi_id)                    ds_filtered = ds_diag_ext.where(                        ds_diag_ext.rgi_id == rgi_id, drop=True)                    if r == 0 or r == 3 or r == 6:                        ax.set_ylabel(variable_axes[v])                    if r >= 6:                        ax.set_xlabel("Time [year]")                    if m == 4:                        ax.plot(ds_filtered["time"], ds_filtered[var] * factors[v],                                label=sample_id, color=colors[model][i], linewidth=2)                    elif i == 0:                        ax.plot(ds_filtered["time"], ds_filtered[var] * factors[v],                                label=sample_id, color=colors[model][i], linewidth=2, alpha=1)                    elif averaged == False:                        ax.plot(ds_filtered["time"], ds_filtered[var] * factors[v],                                label=sample_id, color=colors[model][i], linewidth=2, linestyle="dotted")                    timeseries_data.append(ds_filtered[var] * factors[v])                if len(timeseries_data) > 0:                    stacked_data = np.stack(timeseries_data)                    # print(stacked_data)                    min_values = np.min(stacked_data, axis=0).flatten()                    max_values = np.max(stacked_data, axis=0).flatten()                # print(min_values.size)                ax.fill_between(ds_filtered["time"].values, min_values, max_values,                                color=colors[model][0], alpha=0.3, label=f"{model}, {members[m]-1}-member range")                # if model == "W5E5":                #     opath = os.path.join(                #         wd_path, "summary", f"climate_run_output_baseline_{sample_id}.nc")                #     ds_diag_ext = xr.open_dataset(opath)                #     ds_diag_ext = ds_diag_ext.where(                #         (ds_diag_ext['time'] >= 1985) & (ds_diag_ext['time'] < 2014))                #     ds_diag_ext = ds_diag_ext.where(                #         ds_diag_ext.rgi_id == rgi_id, drop=True)                #     ax.plot(ds_diag_ext["time"], ds_diag_ext[var] * factors[v],                #             label="W5E5.000 baseline", color=colors[model][i], linestyle="-")            axes[r].set_title(rgi_id)        # Adjust the legend        handles, labels = axes[0].get_legend_handles_labels()        fig.legend(handles, labels, loc='lower center',                   bbox_to_anchor=(0.5, -0.1), ncol=5)        plt.tight_layout()        suffix = "member.avg" if averaged else "all.members"        o_folder_data = f"/Users/magaliponds/OneDrive - Vrije Universiteit Brussel/1. VUB/02. Coding/01. IRRMIP/04. Figures/02. OGGM simulations/01. Modelled output/0{v + 1}. {variable_names[v]}/00. Combined"        os.makedirs(o_folder_data, exist_ok=True)        o_file_name = f"{o_folder_data}/1985_2014.{timeframe}.delta.{variable_names[v]}.combined.{suffix}.png"        plt.savefig(o_file_name, bbox_inches='tight')    return# Call the function with appropriate parameters# plot_volume_area_evolution_over_time_combined(True)plot_volume_area_evolution_over_time_combined(False)# %% Cell 8: Plot temperature and precipitation - against W5E5 prepro-4gdirs = workflow.init_glacier_directories(rgi_ids)timeframe = "monthly"colors = {    "W5E5": ["#000000"],  # "#000000"],  # Black    # Darker to lighter shades of purple    "E3SM": ["#785EF0", "#8F7BF1", "#A6A8F2"],    # Darker to lighter shades of pink    "CESM2": ["#DC267F", "#E58A9E", "#F0A2B6", "#F7BCC4"],    # Darker to lighter shades of orange    "CNRM": ["#FE6100", "#FE7D33", "#FE9A66", "#FEB799", "#FECDB5", "#FEF1E1"],    "IPSL-CM6": ["#FFB000"]  # Dark purple to lighter shades}# members=[3,4,6,1,1]members = [1, 1, 1, 1]  # ,1]models = ["E3SM", "CESM2", "CNRM", "IPSL-CM6"]markers = ["o", "x"]# for v,var in enumerate(["prcp", "temp"]):fig, axes = plt.subplots(3, 3, figsize=(22, 8))for r, gdir in enumerate(gdirs):    rgi_id = gdir.rgi_id    row = r // axes.shape[1]    col = r % axes.shape[1]    # ds_diag_base=xr.open_dataset(opath_base)    for m, model in enumerate(models):        for member in range(members[m]):            sample_id = f"{model}.00{member}"            # Create a secondary axis for precipitation            years = np.arange(1985, 2015, 1)            # Open, process and plot the baseline            if m == 0 & member == 0:                axes2 = axes[row, col].twinx()                with xr.open_dataset(gdir.get_filepath('climate_historical')) as ds_diag_base:                    ds_diag_base = ds_diag_base.load()                ds_diag_base = ds_diag_base.sel(                    time=slice("1985-01-01", "2014-12-01"))                ds_diag_base_t = ds_diag_base.temp.resample(time="YS").mean()                ds_diag_base_p = ds_diag_base.prcp.resample(time="YS").sum()                axes2.bar(ds_diag_base_p.time, ds_diag_base_p,  width=pd.Timedelta(                    days=365), edgecolor='black', label="W5E5 Baseline", color="lightgrey", alpha=0.5)                axes[row, col].plot(ds_diag_base_t.time, ds_diag_base_t.values,                                    label='W5E5 Baseline',  color='k', alpha=1)  # marker=markers[v],            with xr.open_dataset(gdir.get_filepath('gcm_data', filesuffix=f'_perturbed_{sample_id}')) as ds_diag_ptb:                ds_diag_ptb = ds_diag_ptb.load()            ds_diag_ptb = ds_diag_ptb.sel(                time=slice("1985-01-01", "2014-12-01"))            ds_diag_ptb_t = ds_diag_ptb.temp.resample(time="YS").mean()            ds_diag_ptb_p = ds_diag_ptb.prcp.resample(time="YS").sum()            # ds_diag_dif=ds_diag_base-ds_diag_ptb            # ds_diag_dif=ds_diag_base-ds_diag_ptb            # axes2.bar(ds_diag_ptb_p.time, ds_diag_ptb_p-ds_diag_base_p, bottom=ds_diag_base_p, width=pd.Timedelta(days=365),label=f'{sample_id}',  edgecolor=colors[model][member], color=colors[model][member], alpha=0.5)#marker=markers[v],            axes2.scatter(ds_diag_ptb_p.time, ds_diag_ptb_p, label=f'{sample_id}',  marker='_',                          color=colors[model][member], alpha=1, linewidth=2, s=150)  # marker=markers[v],            if m == 3 and member == 0:                axes2.bar(ds_diag_base_p.time, ds_diag_base_p,  width=pd.Timedelta(                    days=365), label='W5E5 Baseline',  color='none', alpha=1, edgecolor='black', linewidth=1)  # marker=markers[v],            axes[row, col].plot(ds_diag_ptb_t.time, ds_diag_ptb_t.values,                                label=f'{sample_id}',  color=colors[model][member], alpha=1, linestyle='-')  # marker=markers[v],            axes[row, col].set_title(rgi_id)            axes2.set_title(rgi_id)            # axes2.plot(ds_diag_ptb.time, np.zeros(len(ds_diag_ptb.time)), 'k', linestyle=':')            if r == 3:                axes[row, col].set_ylabel(                    ds_diag_base_t.long_name+" ["+ds_diag_base_t.units+"]")            if r >= 6:                axes[row, col].set_xlabel("Time [year]")            if r == 5:                axes2.set_ylabel(ds_diag_base_p.long_name +                                 "[" + ds_diag_base_p.units+"]")            # axes2.set_ylim(-1000,1000)# Adjust the legendhandles, labels = axes[0, 0].get_legend_handles_labels()handles2, labels2 = axes2.get_legend_handles_labels()handles += handles2labels += labels2fig.legend(handles, labels, loc='lower center',           bbox_to_anchor=(0.5, -0.1), ncol=5)plt.tight_layout()plt.show()o_folder_data = "/Users/magaliponds/OneDrive - Vrije Universiteit Brussel/1. VUB/02. Coding/01. IRRMIP/04. Figures/02. OGGM simulations/00. Processed Climate/Total/"os.makedirs(f"{o_folder_data}/", exist_ok=True)o_file_name = f"{o_folder_data}/total.1985_2014.{timeframe}.processed.climate.data.prepro.difference.png"plt.savefig(o_file_name, bbox_inches='tight')# %%
